Renamed NAWS and Echo detail headers to protocol.hpp
For consistency with new_environ.
